requirements.txt
discord.py
anthropic
openai
asyncio
flask